Changed formatting.

+++ src/util/table.h
@@ -25,7 +25,7 @@ struct Table {
 	}
 
 	Table& operator()(const std::string& s, double n, const char* unit = "") {
-		data_.emplace_back(s, (n >= 100.0 ? std::to_string(std::round(n)) : std::to_string(n)) + unit);
+		data_.emplace_back(s, (n >= 100.0 ? std::to_string((int64_t)std::round(n)) : std::to_string(n)) + unit);
 		max_len_ = std::max(max_len_, s.length());
 		return *this;
 	}
